加入低功耗模式，睡眠和停止模式，二选一，继续优化

+++ Tingzhen_Program/HARDWARE/IWDG/iwdg.c
@@ -6,7 +6,7 @@ IWDG_HandleTypeDef IWDG_Handler; //独立看门狗句柄
 //prer:分频数:IWDG_PRESCALER_4~IWDG_PRESCALER_256
 //rlr:自动重装载值,0~0XFFF.
 //时间计算(大概):Tout=((4*2^prer)*rlr)/32 (ms).
-void IWDG_Init(u8 prer,u16 rlr)
+void IWDG_Init(u8 prer,u32 rlr)
 {
     IWDG_Handler.Instance=IWDG;
     IWDG_Handler.Init.Prescaler=prer;	//设置IWDG分频系数
