+++ Tingzhen_Program/HARDWARE/IWDG/iwdg.h
@@ -3,6 +3,6 @@
 #include "sys.h"
 	
 
-void IWDG_Init(u8 prer,u16 rlr);//初始化IWDG，并使能IWDG
+void IWDG_Init(u8 prer,u32 rlr);//初始化IWDG，并使能IWDG
 void IWDG_Feed(void);//喂狗
 #endif
